refactor : RegisterRequest 패키지 변경
