+++ src/main/java/com/hyerijang/dailypay/auth/AuthenticationController.java
@@ -1,6 +1,5 @@
 package com.hyerijang.dailypay.auth;
 
-import com.hyerijang.dailypay.member.dto.RegisterRequest;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
